HW_3
fix populate DB and SQL command in JdbcMealRepository

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcMealRepository.java
@@ -47,8 +47,8 @@ public class JdbcMealRepository implements MealRepository {
             Number newKey = insertMeal.executeAndReturnKey(map);
             meal.setId(newKey.intValue());
         } else if (namedParameterJdbcTemplate.update(
-                "UPDATE meals SET id=:id, user_id=:userId, date_time=:dateTime, " +
-                        "description=:description, calories=:calories WHERE id=:id", map) == 0) {
+                "UPDATE meals SET date_time=:dateTime, " +
+                        "description=:description, calories=:calories WHERE id=:id AND user_id=:userId", map) == 0) {
             return null;
         }
         return meal;
@@ -56,7 +56,7 @@ public class JdbcMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id, int userId) {
-        jdbcTemplate.update("DELETE FROM meals WHERE id=?", id);
+        jdbcTemplate.update("DELETE FROM meals WHERE id=? AND user_id=:userId", id, userId);
         return true;
     }
 
@@ -68,12 +68,12 @@ public class JdbcMealRepository implements MealRepository {
 
     @Override
     public List<Meal> getAll(int userId) {
-        return jdbcTemplate.query("SELECT * FROM meals WHERE user_id=? ORDER BY date_time", ROW_MAPPER, userId);
+        return jdbcTemplate.query("SELECT * FROM meals WHERE user_id=? ORDER BY date_time DESC", ROW_MAPPER, userId);
     }
 
     @Override
     public List<Meal> getBetweenHalfOpen(LocalDateTime startDateTime, LocalDateTime endDateTime, int userId) {
-        return jdbcTemplate.query("SELECT * FROM meals WHERE date_time BETWEEN? AND?", ROW_MAPPER, startDateTime,
-                endDateTime);
+        return jdbcTemplate.query("SELECT * FROM meals WHERE user_id=? AND date_time BETWEEN ? AND ?", ROW_MAPPER,
+                userId, startDateTime, endDateTime);
     }
 }
